feat(cli): add experimental fullscreen logger type
This is still experimental and needs to be tested across different
terminals and platforms, but imo adds a solid usability improvement
to watch mode commands. Enable it by setting `--logger-type=fullscreen`
or `GARDEN_LOG_LEVEL=fullscreen`.

It is disabled by default, so if you're satisfied that it doesn't mess
with normal operation, I think it might be good to merge this and use
internally for a while to work out any kinks and get some usability
ideas.